requirements.txt
python-dotenv
fastapi
uvicorn
pydantic
pinecone-client
openai
sentence-transformers
requests
flask